moved menu code into own struct
